# api/openapi.yaml
openapi: 3.1.0
info:
  title: Wuling DevOps API
  version: 0.1.0-stage1
  description: |
    Stage 1 of the Wuling DevOps platform.

    What's implemented:
      - Auth (register, login, /me, personal access tokens, SSH public keys)
      - Orgs and Projects (CRUD)
      - Repositories (CRUD + libgit2-backed read APIs)
      - Git smart HTTP (clone/fetch/push) and Git over SSH (default :2222)
      - Issues (CRUD, labels, comments, assignees)
      - Merge Requests (open/diff/merge with FF/merge-commit/squash, comments, reviews)
      - Wiki (per-project Markdown pages backed by a separate Git repo)
      - Insights (per-project activity, contributors, language stats)

    Planned for later stages (URLs reserved, not yet served):
      - Pipelines

    All errors use the JSON envelope:
      { "error": { "code": "<stable_code>", "message": "<human msg>", "details": {} } }
servers:
  - url: http://localhost:8080
    description: Local dev
security:
  - bearerAuth: []
paths:
...
  /api/v1/orgs/{org_slug}/projects/{project_slug}/issues:
    parameters:
      - $ref: "#/components/parameters/OrgSlug"
      - $ref: "#/components/parameters/ProjectSlug"
    get:
      summary: List issues in a project
      parameters:
        - in: query
          name: state
          schema: { type: string, enum: [open, closed] }
        - in: query
          name: label
          schema: { type: string }
          description: Filter by label name (case-insensitive).
        - in: query
          name: assignee
          schema: { type: string }
          description: User UUID or username.
        - in: query
          name: author
          schema: { type: string }
          description: User UUID or username.
        - in: query
          name: search
          schema: { type: string }
          description: Substring match against title and body (case-insensitive).
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  issues:
                    type: array
                    items: { $ref: "#/components/schemas/Issue" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
    post:
      summary: Create an issue
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateIssueRequest" }
      responses:
        "201":
          description: created
          content: { application/json: { schema: { $ref: "#/components/schemas/Issue" } } }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
  /api/v1/orgs/{org_slug}/projects/{project_slug}/issues/{number}:
    parameters:
      - $ref: "#/components/parameters/OrgSlug"
      - $ref: "#/components/parameters/ProjectSlug"
      - $ref: "#/components/parameters/IssueNumber"
    get:
      summary: Get an issue by its per-project number
      responses:
        "200":
          description: ok
          content: { application/json: { schema: { $ref: "#/components/schemas/Issue" } } }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
    patch:
      summary: Edit an issue (author or org owner/admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchIssueRequest" }
      responses:
        "200":
          description: ok
          content: { application/json: { schema: { $ref: "#/components/schemas/Issue" } } }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "403": { $ref: "#/components/responses/ForbiddenError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
    delete:
      summary: Delete an issue (org owner/admin only)
      responses:
        "204": { description: deleted }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "403": { $ref: "#/components/responses/ForbiddenError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
  /api/v1/orgs/{org_slug}/projects/{project_slug}/issues/{number}/comments:
    parameters:
      - $ref: "#/components/parameters/OrgSlug"
      - $ref: "#/components/parameters/ProjectSlug"
      - $ref: "#/components/parameters/IssueNumber"
    get:
      summary: List comments on an issue
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  comments:
                    type: array
                    items: { $ref: "#/components/schemas/IssueComment" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
    post:
      summary: Post a comment on an issue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, minLength: 1, maxLength: 65536 }
      responses:
        "201":
          description: created
          content: { application/json: { schema: { $ref: "#/components/schemas/IssueComment" } } }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
  /api/v1/orgs/{org_slug}/projects/{project_slug}/labels:
    parameters:
      - $ref: "#/components/parameters/OrgSlug"
      - $ref: "#/components/parameters/ProjectSlug"
    get:
      summary: List labels in a project
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  labels:
                    type: array
                    items: { $ref: "#/components/schemas/Label" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
    post:
      summary: Create a label (org owner/admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 64 }
                color:
                  type: string
                  description: 6-character hex (with or without leading "#"). Defaults to 888888.
                description: { type: string, maxLength: 256 }
      responses:
        "201":
          description: created
          content: { application/json: { schema: { $ref: "#/components/schemas/Label" } } }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "403": { $ref: "#/components/responses/ForbiddenError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
  /api/v1/orgs/{org_slug}/projects/{project_slug}/labels/{label_id}:
    parameters:
      - $ref: "#/components/parameters/OrgSlug"
      - $ref: "#/components/parameters/ProjectSlug"
      - in: path
        name: label_id
        required: true
        schema: { type: string, format: uuid }
    delete:
      summary: Delete a label (org owner/admin only)
      responses:
        "204": { description: deleted }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "403": { $ref: "#/components/responses/ForbiddenError" }
        "404": { $ref: "#/components/responses/NotFoundError" }

  # ---------------- merge requests ----------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
      description: |
        Used by Git CLI only. The "password" is either a personal access token
        (preferred — recognised by the `wlpat_` prefix) or the user's account
        password.

        **Security:** credentials are sent on every request, so basic auth
        MUST be used over HTTPS in production — configure server URLs as
        `https://` and terminate TLS in front of the API. For local
        development against `localhost` plain HTTP is acceptable.

        Operators should prefer personal access tokens (`wlpat_…`) over
        account passwords, and consider OAuth/bearer-token flows for
        non-CLI clients.
  parameters:
    OrgSlug:
      in: path
      name: org_slug
      required: true
      schema: { type: string }
    ProjectSlug:
      in: path
      name: project_slug
      required: true
      schema: { type: string }
...
    IssueNumber:
      in: path
      name: number
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
      description: Per-project monotonically increasing issue number (e.g. 1, 2, 3 ...).
...
  responses:
    UnauthorizedError:
      description: missing or invalid credentials
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    ForbiddenError:
      description: caller lacks permission
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFoundError:
      description: resource not found
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    ValidationError:
      description: request validation failed
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - validation
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - already_exists
                - rate_limited
                - unsupported
                - bad_request
                - internal
                - unavailable
            message: { type: string }
            details: { type: object, additionalProperties: true }
...
    UserRef:
      type: object
      description: Lightweight user reference embedded in issue/comment payloads.
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        display_name: { type: string }
    Label:
      type: object
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        name: { type: string }
        color:
          type: string
          description: 6-character hex (no leading "#").
        description: { type: string }
        created_at: { type: string, format: date-time }
    IssueState:
      type: string
      enum: [open, closed]
    Issue:
      type: object
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        number: { type: integer, format: int64 }
        title: { type: string }
        body: { type: string }
        state: { $ref: "#/components/schemas/IssueState" }
        author: { $ref: "#/components/schemas/UserRef" }
        closed_at: { type: string, format: date-time, nullable: true }
        closed_by: { $ref: "#/components/schemas/UserRef", nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        labels:
          type: array
          items: { $ref: "#/components/schemas/Label" }
        assignees:
          type: array
          items: { $ref: "#/components/schemas/UserRef" }
        comment_count: { type: integer, format: int64 }
    CreateIssueRequest:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 256 }
        body: { type: string, maxLength: 65536 }
        labels:
          type: array
          items: { type: string, format: uuid }
        assignees:
          type: array
          items: { type: string, format: uuid }
    PatchIssueRequest:
      type: object
      description: |
        Patch an issue. Omitted fields are left untouched; arrays (labels,
        assignees) are replaced wholesale when present (use [] to clear).
      properties:
        title: { type: string, minLength: 1, maxLength: 256 }
        body: { type: string, maxLength: 65536 }
        state: { $ref: "#/components/schemas/IssueState" }
        labels:
          type: array
          items: { type: string, format: uuid }
        assignees:
          type: array
          items: { type: string, format: uuid }
    IssueComment:
      type: object
      properties:
        id: { type: string, format: uuid }
        issue_id: { type: string, format: uuid }
        author: { $ref: "#/components/schemas/UserRef" }
        body: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...